core network api/client
