+++ commonApp/build.gradle.kts
@@ -14,7 +14,8 @@ kotlin {
                 implementation(project(":core:paging"))
                 implementation(project(":core:todo"))
                 implementation(project(":core:auth"))
-                implementation(project(":core:network"))
+                implementation(project(":core:network:api"))
+                implementation(project(":core:network:client"))
                 implementation(project(":core:store"))
 
                 implementation(project(":feature:home"))
