+++ commonApp/src/commonMain/kotlin/com/stslex/atten/di/AppModules.kt
@@ -3,7 +3,7 @@ package com.stslex.atten.di
 import com.stslex.atten.core.auth.di.ModuleCoreAuth
 import com.stslex.atten.core.core.di.ModuleCore
 import com.stslex.atten.core.database.di.ModuleCoreDatabase
-import com.stslex.atten.core.network.di.ModuleCoreNetwork
+import com.stslex.atten.core.network.client.di.ModuleCoreNetwork
 import com.stslex.atten.core.paging.di.ModuleCorePaging
 import com.stslex.atten.core.todo.di.ModuleCoreToDo
 import com.stslex.atten.core.ui.kit.utils.ModuleCoreUiUtils
