+++ core/network/api/src/commonMain/kotlin/com/stslex/atten/core/network/api/AppHttpClient.kt
@@ -1,0 +1,8 @@
+package com.stslex.atten.core.network.api
+
+import io.ktor.client.HttpClient
+
+interface AppHttpClient {
+
+    val client: HttpClient
+}
