+++ core/network/api/src/commonMain/kotlin/com/stslex/atten/core/network/api/ErrorHandler.kt
@@ -1,0 +1,5 @@
+package com.stslex.atten.core.network.api
+
+import io.ktor.client.plugins.CallRequestExceptionHandler
+
+interface ErrorHandler : CallRequestExceptionHandler
