+++ core/network/api/src/commonMain/kotlin/com/stslex/atten/core/network/api/model/request/AuthGoogleRequest.kt
@@ -1,0 +1,10 @@
+package com.stslex.atten.core.network.api.model.request
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class AuthGoogleRequest(
+    @SerialName("id_token")
+    val idToken: String,
+)
