+++ core/network/api/src/commonMain/kotlin/com/stslex/atten/core/network/api/model/response/LoginOkResponse.kt
@@ -1,0 +1,16 @@
+package com.stslex.atten.core.network.api.model.response
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class LoginOkResponse(
+    @SerialName("uuid")
+    val uuid: String,
+    @SerialName("email")
+    val email: String,
+    @SerialName("access_token")
+    val accessToken: String,
+    @SerialName("refresh_token")
+    val refreshToken: String
+)
