+++ core/network/client/src/commonMain/kotlin/com/stslex/atten/core/network/client/di/ModuleCoreNetwork.kt
@@ -1,0 +1,10 @@
+package com.stslex.atten.core.network.client.di
+
+import org.koin.core.annotation.ComponentScan
+import org.koin.core.annotation.Module
+import org.koin.core.annotation.Singleton
+
+@Module
+@ComponentScan("com.stslex.atten.core.network.client")
+@Singleton
+class ModuleCoreNetwork
