--- core/network/src/commonMain/kotlin/com/stslex/atten/core/network/client/AppHttpClient.kt
@@ -1,9 +1,0 @@
-package com.stslex.atten.core.network.client
-
-import io.ktor.client.HttpClient
-
-interface AppHttpClient {
-
-    val client: HttpClient
-}
-
