--- core/network/src/commonMain/kotlin/com/stslex/atten/core/network/di/ModuleCoreNetwork.kt
@@ -1,10 +1,0 @@
-package com.stslex.atten.core.network.di
-
-import org.koin.core.annotation.ComponentScan
-import org.koin.core.annotation.Module
-import org.koin.core.annotation.Singleton
-
-@Module
-@ComponentScan("com.stslex.atten.core.network")
-@Singleton
-class ModuleCoreNetwork
