--- core/network/src/commonMain/kotlin/com/stslex/atten/core/network/error_handler/ErrorHandler.kt
@@ -1,5 +1,0 @@
-package com.stslex.atten.core.network.error_handler
-
-import io.ktor.client.plugins.CallRequestExceptionHandler
-
-interface ErrorHandler : CallRequestExceptionHandler
