--- core/network/src/commonMain/kotlin/com/stslex/atten/core/network/model/TokenResponseModel.kt
@@ -1,16 +1,0 @@
-package com.stslex.atten.core.network.model
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class TokenResponseModel(
-    @SerialName("uuid")
-    val uuid: String,
-    @SerialName("email")
-    val email: String,
-    @SerialName("access_token")
-    val accessToken: String,
-    @SerialName("refresh_token")
-    val refreshToken: String
-)
