+++ settings.gradle.kts
@@ -26,7 +26,8 @@ include(":core:ui:mvi")
 include(":core:database")
 include(":core:paging")
 include(":core:todo")
-include(":core:network")
+include(":core:network:api")
+include(":core:network:client")
 include(":core:auth")
 include(":core:store")
